requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
accelerate
bitsandbytes
torch

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
sentence-transformers
chromadb
